fix: an `f80` literal is encoded in eighty bits
  before:  store f80 fconst:f80:0x3ff0000000000000    (f64's 1.0)
  after:   store f80 fconst:f80:0x3fff8000000000000000 (what gcc stores)

`Const::Float`'s payload widens from `u64` to `u128`, because eighty bits do not fit in
sixty-four -- the representation change my section 9 note said was unnecessary. `f32` and `f64`
print identically either way, so no golden moved.

`x87_bits` does the re-encoding, and the special cases are why it is not three lines:

  - a **zero** has a zero exponent and a zero significand in both formats, and rebiasing it
    would produce a denormal with the integer bit set, which x87 calls invalid
  - an **infinity or NaN** has the all-ones exponent in both, so the exponent is mapped rather
    than shifted and a NaN's payload rides along in the fraction
  - a **normal** number is rebiased by 15360, gets the explicit integer bit x87 has and `f64`
    does not, and its fraction shifts left by eleven

The expected patterns come from gcc's object bytes rather than from my arithmetic, because
deriving them is the step that produced the bug in the first place.

**Two process notes worth keeping.**

`cargo test --workspace | grep "^test .* FAILED"` reported nothing while the build was broken:
five test files construct `Const::Float` with `u64`, so nothing ran. **A build failure looks
exactly like zero failures to that grep** -- `--no-run` first, or count errors, before believing
a clean sweep.

The declared gap is unchanged: the engine still models no `f80` arithmetic and still says so.
What changed is that when it eventually does, the bits it starts from are the right ones.

1363 passing, 0 failed, all gates green.

+++ crates/chiero-check/tests/undefined_arithmetic.rs
@@ -639,7 +639,7 @@ fn a_float_cast_overflow_is_reported() {
                     from: CTy::Float(FloatKind::F64),
                     a: Operand::Const(Const::Float(
                         FloatKind::F64,
-                        (-4_294_905_087.0f64).to_bits(),
+                        u128::from((-4_294_905_087.0f64).to_bits()),
                     )),
                 },
             },
@@ -664,7 +664,7 @@ fn a_float_cast_overflow_is_reported() {
                     kind: CastKind::FpToSi,
                     to: CTy::Int(32),
                     from: CTy::Float(FloatKind::F64),
-                    a: Operand::Const(Const::Float(FloatKind::F64, 2.7f64.to_bits())),
+                    a: Operand::Const(Const::Float(FloatKind::F64, u128::from(2.7f64.to_bits()))),
                 },
             },
             10,
@@ -721,7 +721,7 @@ fn every_ub_kind_is_spelled_as_a_slug() {
                     from: CTy::Float(FloatKind::F64),
                     a: Operand::Const(Const::Float(
                         FloatKind::F64,
-                        (-4_294_905_087.0f64).to_bits(),
+                        u128::from((-4_294_905_087.0f64).to_bits()),
                     )),
                 },
             },
